requirements.txt
beautifulsoup4>=4.12.0
certifi>=2023.7.22
charset-normalizer>=3.0.0
contourpy>=1.2.0
cycler>=0.11.0
fonttools>=4.40.0
httpx>=0.25.0
idna>=3.4
joblib>=1.3.0
kiwisolver>=1.4.0
matplotlib>=3.8.0
numpy>=1.24.0
packaging>=23.0
pandas>=2.1.0
pillow>=10.0.0
pyparsing>=3.1.0
python-dateutil>=2.8.2
python-dotenv>=1.0.0
pytz>=2023.3
regex>=2023.6.3
requests>=2.31.0
scikit-learn>=1.3.0
scipy>=1.11.0
seaborn>=0.12.2
six>=1.16.0
soupsieve>=2.4.0
tavily-python>=0.3.0
threadpoolctl>=3.2.0
tiktoken>=0.5.0
typing_extensions>=4.7.0
tzdata>=2023.3
urllib3>=2.0.0
